app/vlinsert/loki: optimize parsing loki protobuf in the way similar to the optimization applied for opentelemetry protobuf in the commit 4ffb74448d270c1366d941a79b88bc3b8d6236fa
The general idea is to parse log entries one-by-one and immediately push them to logMessageProcessor
instead of constructing intermediate data structures.

This reduces memory usage because there is no need in holding the intermediate data structures in memory.
This also reduces CPU usage a bit.

Also make sure that every log entry in the same protobuf message is assigned with an unique timestamp
if the log entry misses the timestamp. Previously all the log entries from the same protobuf message
were assigned the same timestamp. This could result in the loss of the original order of the log messages.